+++ pages/mypage/index.tsx
@@ -1,244 +1,17 @@
-import COLOR from '@constants/colors';
-import FONT from '@constants/fonts';
-import Image from 'next/image';
-import React, { useState } from 'react';
-import { styled } from 'styled-components';
-import { AiFillHeart } from 'react-icons/ai';
-import { IPlace } from '@@types/facility';
-import { BsArrowLeft } from 'react-icons/bs';
-import { useRouter } from 'next/router';
-
-const PlaceTypeDic = {
-  museum: '박물관',
-  artGallery: '미술관',
-  exhibition: '전시회'
-};
+import ProfileSection from '@mypage/ProfileSection';
+import BadgeSection from '@mypage/BadgeSection';
+import WishPlaceSection from '@mypage/WishPlaceSection';
+import MypageHeader from '@mypage/MypageHeader';
 
 const Mypage = () => {
-  const [wishPlaceList, setWishPlaceList] = useState<IPlace[]>([
-    {
-      id: '0',
-      name: '국립 고궁 박물관',
-      type: 'museum',
-      address: '서울 종로구 세종로',
-      opening_time: '10:00'
-    },
-    {
-      id: '1',
-      name: '국립 고궁 박물관',
-      type: 'museum',
-      address: '서울 종로구 세종로',
-      opening_time: '10:00'
-    }
-  ]);
-
-  const [badgeList, setBadgetList] = useState([
-    {
-      id: '0',
-      imageSrc: '/images/elevator.svg',
-      name: '어디든지 완전정복'
-    },
-    {
-      id: '1',
-      imageSrc: '/images/elevator.svg',
-      name: '어디든지 완전정복'
-    },
-    {
-      id: '2',
-      imageSrc: '/images/elevator.svg',
-      name: '어디든지 완전정복'
-    },
-    {
-      id: '3',
-      imageSrc: '/images/elevator.svg',
-      name: '어디든지 완전정복'
-    },
-    {
-      id: '4',
-      imageSrc: '/images/elevator.svg',
-      name: '어디든지 완전정복'
-    }
-  ]);
-
-  const router = useRouter();
-
   return (
     <div>
-      <MypageHeader>
-        <CursorWrapper>
-          <BsArrowLeft color={COLOR.GREY} size={25} onClick={() => router.back()} />
-        </CursorWrapper>
-        <HeaderText style={FONT.BODY1}>마이페이지</HeaderText>
-        <Blank>빈칸</Blank>
-      </MypageHeader>
-
-      <ProfileSection>
-        <ProfileWrapper>
-          <Image src='/images/elevator.svg' alt='' width={50} height={50} />
-          <UserIDText style={FONT.HEADLINE2}>사용자 아이디 님</UserIDText>
-        </ProfileWrapper>
-        <LogoutText style={FONT.CAPTION}>로그아웃</LogoutText>
-      </ProfileSection>
-
-      <BadgeSection>
-        <TitleText style={FONT.HEADLINE1}>나의 뱃지</TitleText>
-        <BadgeContainer>
-          {badgeList.map((badge) => (
-            <BadgeWrapper key={badge.id}>
-              {/* 성공 했을 때와 성공안했을 때 분기처리 */}
-              {/* <Image src='/images/elevator.svg' alt='' width={80} height={80} /> */}
-              <NotCompletedBadge />
-              <BadgeName style={FONT.BODY2}>어디든지 완전정복</BadgeName>
-            </BadgeWrapper>
-          ))}
-        </BadgeContainer>
-      </BadgeSection>
-
-      <WishPlaceSection>
-        <TitleText style={FONT.HEADLINE1}>찜한 시설</TitleText>
-
-        {wishPlaceList.map((place) => (
-          <WishPlaceWrapper key={place.id}>
-            <WishPlaceImage>
-              <Image src='/images/elevator.svg' alt='' width={80} height={80} />
-            </WishPlaceImage>
-            <WishPlaceInfo>
-              <PlaceName style={FONT.HEADLINE2}>{place.name}</PlaceName>
-              <PlaceType style={FONT.BODY2} $type={place.type}>
-                {/* {PlaceTypeDic[place.type]} */}
-                {PlaceTypeDic['museum']}
-              </PlaceType>
-              <PlaceLocation style={FONT.BODY2}>{place.address}</PlaceLocation>
-            </WishPlaceInfo>
-            <HeartButton>
-              <AiFillHeart color='red' size={25} />
-            </HeartButton>
-          </WishPlaceWrapper>
-        ))}
-      </WishPlaceSection>
+      <MypageHeader />
+      <ProfileSection />
+      <BadgeSection />
+      <WishPlaceSection />
     </div>
   );
 };
-
-const MypageHeader = styled.div`
-  display: flex;
-  justify-content: space-between;
-  padding: 30px;
-`;
-
-const CursorWrapper = styled.div`
-  cursor: pointer;
-`;
-
-const HeaderText = styled.div``;
-
-const Blank = styled.div`
-  color: white;
-`;
-
-const ProfileSection = styled.div`
-  display: flex;
-  justify-content: space-between;
-  align-items: center;
-  padding: 30px;
-`;
-
-const ProfileWrapper = styled.div`
-  display: flex;
-  align-items: center;
-  gap: 20px;
-`;
-
-const UserIDText = styled.div``;
-
-const LogoutText = styled.div`
-  color: ${COLOR.GREY};
-  cursor: pointer;
-`;
-
-const PlaceName = styled.div``;
-
-const PlaceLocation = styled.div`
-  color: ${COLOR.GREY};
-`;
-
-interface PlaceTypeProps {
-  $type: string;
-}
-
-type ObjType = {
-  [index: string]: string;
-};
-
-const TYPE_TO_COLOR: ObjType = {
-  museum: COLOR.ORANGE,
-  artGallery: COLOR.GREEN,
-  exhibition: COLOR.RED
-};
-
-const PlaceType = styled.div<PlaceTypeProps>`
-  color: ${(props) => TYPE_TO_COLOR[props.$type]};
-`;
-
-const BadgeContainer = styled.div`
-  display: flex;
-  justify-content: center;
-  flex-wrap: wrap;
-  gap: 30px;
-`;
-
-const BadgeSection = styled.div``;
-
-const BadgeWrapper = styled.div`
-  display: flex;
-  flex-direction: column;
-  align-items: center;
-  gap: 10px;
-`;
-
-const NotCompletedBadge = styled.div`
-  background-color: ${COLOR.LINE};
-  width: 80px;
-  height: 80px;
-  border-radius: 16px;
-`;
-
-const WishPlaceSection = styled.div``;
-
-const TitleText = styled.div`
-  padding: 20px;
-`;
-
-const BadgeName = styled.div`
-  width: 60%;
-  text-align: center;
-  display: flex;
-  flex-wrap: wrap;
-`;
-
-const WishPlaceWrapper = styled.div`
-  display: flex;
-  justify-content: space-between;
-  align-items: center;
-  margin-bottom: 20px;
-`;
-
-const WishPlaceImage = styled.div`
-  flex-basis: 20%;
-`;
-
-const WishPlaceInfo = styled.div`
-  flex-basis: 60%;
-  padding-left: 20px;
-  display: flex;
-  flex-direction: column;
-  gap: 3px;
-`;
-
-const HeartButton = styled.div`
-  flex-basis: 20%;
-  display: flex;
-  justify-content: center;
-`;
 
 export default Mypage;
